requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
PyMuPDF 
langchain-text-splitters
langchain-openai
langchain-ollama
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
uvicorn
fastapi
langserve[all]